lavazza-coffee-agent/requirements.txt
# LangGraph + LLM
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0

# Database
pymongo>=4.0.0
qdrant-client>=1.9.0

# Data
pandas>=2.0.0
numpy>=1.24.0
httpx>=0.27.0
yfinance>=0.2.0
faostat>=1.0.0

# Dashboard
streamlit>=1.35.0
plotly>=5.22.0

# PDF (monthly reports)
weasyprint>=62.0

# Utils
python-dotenv>=1.0.0
openai>=1.0.0
